Refactor models to use uint64 for primary keys
- Updated primary key type from uint to uint64 in multiple models including Activity, Attendance, Class, Exam, Expense, Item, Leave, Lesson, Log, Notification, Parent, Payment, PlanningSlot, Planning, Remediation, Remuneration, Role, SchoolYear, Staff, StudentClass, Student, Synchronization, Task, TeacherLesson, Teachers, TransactionHistory, and User.
- Adjusted corresponding functions to accept uint64 for ID parameters in CRUD operations.

+++ models/activity.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Activity struct {
-	ID          uint      `gorm:"primaryKey"`
+	ID          uint64      `gorm:"primaryKey"`
 	Title       string    `gorm:"size:100;not null"`
 	Description string    `gorm:"type:text"`
 	Type        string    `gorm:"size:50"`
@@ -29,7 +29,7 @@ func CreateActivity(db *gorm.DB, activity *Activity) error {
 }
 
 // GetActivityByID récupère une activité par ID
-func GetActivityByID(db *gorm.DB, id uint) (*Activity, error) {
+func GetActivityByID(db *gorm.DB, id uint64) (*Activity, error) {
 	var activity Activity
 	err := db.First(&activity, id).Error
 	if err != nil {
@@ -51,7 +51,7 @@ func UpdateActivity(db *gorm.DB, activity *Activity) error {
 }
 
 // DeleteActivity supprime une activité (soft delete)
-func DeleteActivity(db *gorm.DB, id uint) error {
+func DeleteActivity(db *gorm.DB, id uint64) error {
 	return db.Delete(&Activity{}, id).Error
 }
