+++ models/roles.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Role struct {
-	ID        uint   `gorm:"primaryKey"`
+	ID        uint64   `gorm:"primaryKey"`
 	Title     string `gorm:"size:100;not null;unique"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
@@ -42,7 +42,7 @@ func UpdateRole(db *gorm.DB, role *Role) error {
 }
 
 // Supprime une Role (soft delete)
-func DeleteRole(db *gorm.DB, id uint) error {
+func DeleteRole(db *gorm.DB, id uint64) error {
 	return db.Delete(&Role{}, id).Error
 }
